+++ R/zzz.R
@@ -1,3 +1,7 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage("Welcome to GPStat!, part of the Daily Measure software suite.")
   shiny::addResourcePath(